% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_delete_report.R
\name{delete_report}
\alias{delete_report}
\title{Delete the BLAST Report Files}
\usage{
delete_report()
}
\value{
No return value. Deletes files as a side effect.
}
\description{
Deletes both the \code{blast_history_report.Rmd} and \code{blast_history_report.html} files from the
working directory.
}
\details{
This function is useful if you want to start a new BLAST analysis and clear out the
previously generated report. It is non-reversible and should be used with caution.
}
\examples{
\dontrun{
# Delete the current BLAST report and start fresh
delete_report()
}

}
