% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_delete_report.R
\name{generate_report}
\alias{generate_report}
\title{Generate the BLAST HTML Report}
\usage{
generate_report()
}
\value{
No return value. An HTML file \code{blast_history_report.html} is created as a side effect.

#' @importFrom rmarkdown render
}
\description{
Uses \code{rmarkdown::render()} to knit the \code{blast_history_report.Rmd} file into an HTML report.
This function must be called by the user after running BLAST-related functions if they want
to generate a formatted report of their analysis.
}
\details{
The function assumes that a file named \code{blast_history_report.Rmd} exists in the current working
directory. This file is expected to have been incrementally updated by internal logging functions
such as \code{reporter_function()}.
}
\examples{

\dontrun{
# Generate the BLAST HTML report
generate_report()
}
}
